+++ package.json
@@ -51,6 +51,7 @@
     "build": "npm run compile && NODE_OPTIONS='--max-old-space-size=4096' npm run dist",
     "bundlesize": "bundlesize",
     "check-commit": "node ./scripts/check-commit",
+    "check-ts-demo": "node ./scripts/check-ts-demo",
     "prestart": "npm run version",
     "precompile": "npm run version",
     "pretest": "npm run version",
